sigsegv error was solved by update tensorflow-addons==0.10 / move photometric loss to new file loss_util.py
