+++ model/loss_and_metric/losses.py
@@ -7,6 +7,7 @@ import utils.util_funcs as uf
 from utils.util_class import WrongInputException
 import utils.convert_pose as cp
 from utils.decorators import shape_check
+from model.loss_and_metric.loss_util import photometric_loss_l1, photometric_loss_ssim
 
 
 class TotalLoss:
@@ -124,80 +125,12 @@ class PhotometricLossMultiScale(PhotometricLoss):
         losses = []
         for i, (synt_target, orig_target) in enumerate(zip(synth_target_ms, original_target_ms)):
             loss = layers.Lambda(lambda inputs: self.photometric_loss(inputs[0], inputs[1]),
-                                 name=f"photo_loss_{i}")([synt_target, orig_target])
+                                 name=f"photo_loss_{i}" + self.key_suffix)([synt_target, orig_target])
             losses.append(loss)
         # losses: list of [batch, num_src] -> after stack: [batch, num_src, num_scale] -> after sum: [batch]
         batch_loss = layers.Lambda(lambda data: tf.reduce_sum(tf.stack(data, axis=2), axis=[1, 2]),
-                                   name="photo_loss_sum")(losses)
+                                   name="photo_loss_sum" + self.key_suffix)(losses)
         return batch_loss
-
-
-@shape_check
-def photometric_loss_l1(synt_target, orig_target):
-    """
-    :param synt_target: scaled synthesized target image [batch, num_src, height/scale, width/scale, 3]
-    :param orig_target: scaled original target image [batch, height/scale, width/scale, 3]
-    :return: photo_loss [batch, num_src]
-    """
-    orig_target = tf.expand_dims(orig_target, axis=1)
-    # create mask to ignore black region
-    synt_target_gray = tf.reduce_mean(synt_target, axis=-1, keepdims=True)
-    error_mask = tf.equal(synt_target_gray, 0)
-
-    # orig_target: [batch, 1, height/scale, width/scale, 3]
-    # axis=1 broadcasted in subtraction
-    # photo_error: [batch, num_src, height/scale, width/scale, 3]
-    photo_error = tf.abs(synt_target - orig_target)
-    photo_error = tf.where(error_mask, tf.constant(0, dtype=tf.float32), photo_error)
-    # average over image dimensions (h, w, c)
-    photo_loss = tf.reduce_mean(photo_error, axis=[2, 3, 4])
-    return photo_loss
-
-
-@shape_check
-def photometric_loss_ssim(synt_target, orig_target):
-    """
-    :param synt_target: scaled synthesized target image [batch, num_src, height/scale, width/scale, 3]
-    :param orig_target: scaled original target image [batch, height/scale, width/scale, 3]
-    :return: photo_loss [batch, num_src]
-    """
-    num_src = synt_target.get_shape().as_list()[1]
-    orig_target = tf.expand_dims(orig_target, axis=1)
-    orig_target = tf.tile(orig_target, [1, num_src, 1, 1, 1])
-    # create mask to ignore black region
-    synt_target_gray = tf.reduce_mean(synt_target, axis=-1, keepdims=True)
-    error_mask = tf.equal(synt_target_gray, 0)
-
-    x = orig_target     # [batch, num_src, height/scale, width/scale, 3]
-    y = synt_target     # [batch, num_src, height/scale, width/scale, 3]
-    c1 = 0.01 ** 2
-    c2 = 0.03 ** 2
-    ksize = [1, 3, 3]
-
-    # TODO IMPORTANT!
-    #   tf.nn.avg_pool results in error like ['NoneType' object has no attribute 'decode']
-    #   when training model with gradient tape in eager mode,
-    #   but no error in graph mode by @tf.function
-    #   Instead, tf.keras.layers.AveragePooling3D results in NO error in BOTH modes
-    # mu_x, mu_y: [batch, num_src, height/scale, width/scale, 3]
-    average_pool = tf.keras.layers.AveragePooling3D(pool_size=ksize, strides=1, padding="SAME")
-    mu_x = average_pool(x)
-    mu_y = average_pool(y)
-    # mu_x = tf.nn.avg_pool(x, ksize=ksize, strides=1, padding='SAME')
-    # mu_y = tf.nn.avg_pool(y, ksize=ksize, strides=1, padding='SAME')
-
-    sigma_x = average_pool(x ** 2) - mu_x ** 2
-    sigma_y = average_pool(y ** 2) - mu_y ** 2
-    sigma_xy = average_pool(x * y) - mu_x * mu_y
-
-    ssim_n = (2 * mu_x * mu_y + c1) * (2 * sigma_xy + c2)
-    ssim_d = (mu_x ** 2 + mu_y ** 2 + c1) * (sigma_x + sigma_y + c2)
-    ssim = ssim_n / ssim_d
-    ssim = tf.clip_by_value((1 - ssim) / 2, 0, 1)
-    ssim = tf.where(error_mask, tf.constant(0, dtype=tf.float32), ssim)
-    # average over image dimensions (h, w, c)
-    ssim = tf.reduce_mean(ssim, axis=[2, 3, 4])
-    return ssim
 
 
 class SmoothenessLossMultiScale(LossBase):
@@ -271,6 +204,7 @@ class StereoDepthLoss(PhotometricLoss):
                                                    target_depth_ms=predictions["depth_ms"],
                                                    pose_t2s=tf.linalg.inv(features["stereo_T_LR"]),
                                                    intrinsic=features["intrinsic"])
+        # synthesize right image from left image
         loss_right, _ = self.stereo_synthesize_loss(source_img=augm_data["target"],
                                                     target_ms=augm_data["target_ms_R"],
                                                     target_depth_ms=predictions["depth_ms_R"],
@@ -338,11 +272,11 @@ class FlowWarpLossMultiScale(PhotometricLoss):
         losses = []
         for i, (warp_target, orig_target) in enumerate(zip(warped_target_ms, flow_target_ms)):
             loss = layers.Lambda(lambda inputs: self.photometric_loss(inputs[0], inputs[1]),
-                                 name=f"flow_warp_loss_{i}")([warp_target, orig_target])
+                                 name=f"flow_warp_loss_{i}" + self.key_suffix)([warp_target, orig_target])
             losses.append(loss)
         # losses: list of [batch, num_src] -> after stack: [batch, num_src, num_scale] -> after sum: [batch]
         batch_loss = layers.Lambda(lambda data: tf.reduce_sum(tf.stack(data, axis=2), axis=[1, 2]),
-                                   name="flow_warp_loss_sum")(losses)
+                                   name="flow_warp_loss_sum" + self.key_suffix)(losses)
         return batch_loss
