+++ requirements.txt
@@ -45,7 +45,7 @@ six==1.13.0
 tensorboard==2.2.2
 tensorboard-plugin-wit==1.7.0
 tensorflow==2.2.0
-tensorflow-addons==0.8.3
+tensorflow-addons==0.10.0
 tensorflow-estimator==2.2.0
 tensorflow-gpu==2.0.0
 termcolor==1.1.0
